# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Be Found Online APis",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/items/categories": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Get all categories",
                "description": "Retrieve a list of all categories",
                "operationId": "232f80091b1d3ab6ded441db60a7d208",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Language header (ar or en)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ar",
                            "example": "en"
                        }
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "Comma separated list of fields to be returned",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[slug]",
                        "in": "query",
                        "description": "Filter by slug (exact or partial depending on your filter)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[name]",
                        "in": "query",
                        "description": "Filter by translated name (LIKE)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[id]",
                        "in": "query",
                        "description": "Filter by id (exact)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "filter[status]",
                        "in": "query",
                        "description": "Filter by status (exact)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[parent_id]",
                        "in": "query",
                        "description": "Filter by parent_id (exact)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort fields. Use leading '-' for DESC. Examples: name, -name, slug, -id",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "asc"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categories retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Categories retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Category Name"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Category Description"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/items/categories/{category}": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Get a single category",
                "description": "Retrieve details of a specific category by ID",
                "operationId": "2411a1a58579643c0225c4137fc71128",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Language header (ar or en)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ar",
                            "example": "en"
                        }
                    },
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "fields",
                        "in": "path",
                        "description": "Comma separated list of fields to be returned",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "id,name,description"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Category retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Category Name"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Category Description"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}